Migrate to module layout
Most functionality is now in the zeekscript module (with parallels to zkg). It
uses setuptools for packaging, with an extended build command that knows to
rebuild the bindings as needed.

Also bumps the tree-sitter-zeek submodule to a new version that now has the
generated parser sources in its own submodule. That ways people don't need
tree-sitter itself installed to get started with this Python package.