Add zeek-format toplevel script
Most people will want to use this package to format scripts, so this simplifies
that use case. The command-line options are identical to that of "zeek-script
format" since the same argparse code creates them. I considered symlinks and a
multi-call script but this looks less portable, and symlinks aren't well
supported by setup.py to begin with.