Differentiate named and unnamed TS nodes
A TS node's type string captures both named and unnamed nodes -- for example, it
may be 'expr' to indicate an expression rule, or ';' to indicate a token. This
can cause ambiguity, for example when a token happens to have the value of a
grammar rule (say "expr"). To resolve this, this commit introduces Node.name()
and Node.token() to distinguish the two, and uses the relevant variant
throughout the Formatter code.