Support recursive descent on directories when formatting scripts
This adds the --recursive|-r flag to zeek-format and "zeek-script format". When
set, the commands process directory names recursively, formatting any *.zeek
file in them. This requires --inplace and will error out when not given.

In the absence of --recursive, directories are now properly recognized, warned
about, and skipped.

Includes testcases. Looks like there's still a line terminator problem on
Windows, which we'll fix separately. These tests are disabled on Windows for
now.