When building up parse tree, track non-error nodes separately
The Node class now has a new member nonerr_children, a subset of the existing
children list that omits ERROR nodes, which can occur anywhere in the tree. This
unpredictability breaks formatter code that expects child nodes to be of certain
types, which is another cause of swalled script content.