Improve linebreaks in enum declarations and initializer lists
We now support heuristics for when to linebreak such constructs. In essence, we
honor existing newlines and take them to mean that individual elements should be
on individual lines. Without any newlines, we preserve single-line format. For
initializer lists we use the same approach, plus the requirement that all
expressions in the list are atomic (IDs or constants) -- otherwise more complex
initializers get linewrapped in the output stage, which looks bad.

Includes updated testcase.

Fixes #3