Reflect removal of init tokens from the grammar
The grammar change unifies initializer handling in expressions, removing the
"init" token from the grammer in the process. This allows some simplifications
in the formatter, moving it to expression handling.

There's a minor behavioral difference in this change, removing somewhat
over-eager heuristics for breaking expression lists across lines that previously
only applied to { ... } initializers in declarations. This might well get
revised in the future.