Add pre-commit hook for running pylint
This is currently running with a similar setup as the existing test
`tests/test_pylint.py` so that we only fail on pylint errors. Since this
pre-commit hook is using a fixed version of pylint we will address
pylint messages in subsequent commits so we can fail for any new
messages.