Drop manually implemented test runner in favor of pytest.
We previously allowed running each test file, and also provided a
wrapper `tests/test_all.py` which would invoke all tests. This
was effectively a one-off implementation of a test runner.

This patch removes code for this runner implementation and instead
delegates completely to pytest which we add as a new CI dependency. Our
tests follow the usual naming convention for tests so pytest can already
discover them. With pytest we get the same functionality as running all
tests (on the toplevel: `pytest`), and running individual test files
(`pytest test/<TEST>`). We also gain the ability to invoke individual
test cases.