Remove `pyproject.toml` `[project]` section
It seems newer versions of the installation toolchain require either the
majority of keywords from `setup.py` declared in `pyproject.toml`, or
none at all (in that case there should be no `[project]` section at
all). As written the project configuration is rejected and it is
impossible to install this package.

Since we use a build controlled by `setup.py` remove the `[project]`
section in `pyproject.toml`.