Add test for snapshot testing
We are currently using a huge file `tests/data/test1.zeek` for manual
"snapshot testing". The idea is that one would add samples to the file
and then manually update the baseline `tests/data/test1.zeek.out`. As
this file grows this manual process becomes very cumbersome.
Additionally when the test fails it can be very hard to identify which
parts of the baseline differed.

This patch introduces a dedicated test `test_samples` which automates
this. The idea is that one would add new compact samples to
`test/samples` which the test discovers automatically. When the test is
run equality assertions are automatically routed to checks against
managed snapshots which can be updated with `pytest --snapshot-update`.
This should reduce overhead from adding/updating snapshots which should
make it easier to add coverage. We will migrate the existing tests over
in follow-up commits.