Simplify baseline names in snapshot testing
We previously would include a running number for the specific test
parameterization in the snapshot for each sample test. This meant that
tests could get renamed whenever a new sample was added.

With this patch we now override the default name enumerating each
parameterization and instead set it to the name of the sample a test
runs on. With that snapshots should be much more stable.