Add type hints
This is largely mechanical and only happening in signatures, but
sometimes also required new types, e.g., to properly type mixin
classes. In a few places I restructured code so it is easier for type
checkers to detect the actual code, or encoded assumptions we already
made (e.g., that a function never returned `None`).